requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

pip==23.1.2
azure-functions
azure-functions-durable>=1.2.4
azure-storage-blob==12.8.0
azure-identity
pandas
openpyxl
openai==0.28.1
AzureOpenAI

requests
beautifulsoup4